refactor(Core/DB): align realm=-1 handling with RBAC convention
Store -1 literally in the DB (signed INT column, matching RBAC) and
resolve it to realm.Id.Realm at load/store time via ResolveRealm().
Drops the REALM_ID_ALL uint32 sentinel and the fallback map lookup it
required. All realmId parameters now use int32 throughout the Mgr API.

+++ src/mod_quest_helper_script.cpp
@@ -57,7 +57,7 @@ public:
             return;
 
         uint32 questId = quest->GetQuestId();
-        uint32 realmId = realm.Id.Realm;
+        int32  realmId = static_cast<int32>(realm.Id.Realm);
 
         LOG_DEBUG("module.quest_helper", "QuestHelper: Player {} accepted quest {} ({})",
             player->GetName(), quest->GetTitle(), questId);
